Work on #5 * Make sure empty data will not be added
