+++ src/tobi.backfrak.de/internal/gpxbl/GpsReader_test.go
@@ -57,23 +57,6 @@ func TestReadNotGPX(t *testing.T) {
 	fmt.Println(gpx.Name)
 }
 
-func TestGpxFileErrorStruct(t *testing.T) {
-
-	path := "/some/sample/path"
-	err := newGpxFileError(path)
-	checkGpxFileError(err, path, t)
-}
-
-func checkGpxFileError(err *GpxFileError, path string, t *testing.T) {
-	if strings.Contains(err.Error(), path) == false {
-		t.Errorf("The error messaage of GpxFileError does not contain the expected Path")
-	}
-
-	if err.File != path {
-		t.Errorf("The GpxFileError.File does not match the expected value")
-	}
-}
-
 func TestReadValidMultiSegmentGPX(t *testing.T) {
 	gpx, err := ReadGPX(testhelper.GetValidGPX("02.gpx"))
 	if err != nil {
